Rename postgres sample app to pg
